Fix information about accessibility
